requirements.txt
google-genai
pydantic>=2.0
fastapi
uvicorn[standard]
python-dotenv